Move new helpers into utils.ts and add proper types for error
